Update CI config and README for older versions of cmake
Although presets arrived with 3.19, conan v2 uses functionality that was
release with 3.23. The current ubuntu LTS, which is used on circleCI,
has 3.22 so we can't use the presets created by conan in that config.

The presets do make config and build simpler, but we must assume that not
all people are on a new enough version of cmake so the README needs to
have instructions for those users.