Move the conan step into cmake to simply configuration
New cmake file wraps up conan functionality and is called from the top
level cmake file. The setting of the toolchain file is done in the top
level file incase we decide to not do the conan step from within cmake.
If so, we can just delete the new file and the call to it.

The README and CI have also been updated to reflect these changes.