Migrated to conan v2 (#16)
* Update conanfile to work with version 2

The way generators are handled has changed so update that section

Changes in conan 2 mean packages *must* be all lowercae and the
Qt-Color-Widget package has been updated so this project needs to track
the change.

Update option string to work with conan v2.

* Update top level CMakeLists file and README with config instructions

Conan v2 has changed the way it interacts with CMake. At the moment, you
need to clone the cmake-conan repo separately and point to it's .cmake
file when configuring your project.

Perhaps the cmake-conan repo should be added as a submodule of the main
project?

* Specify all the packages that are used

This is needed to get the project compiling. Perhaps the method of
calling conan1 automagically found and included everything. This does
not seem to be the case with conan2 so we have to manaully add.

* Add the cstdint include for uint16_t

This was required to compile with gcc 13.1.1

* Ignore more autogenerated files

Conan now creates a preset file that can be used with newer versions of
cmake.

IDE like feature with some editors (e.g. neovim) work best when you get
cmake to create a `compile_commands.json` file that can be used by lsp.
This then creates a `.cache` file where is does it's work. We shouldn't
track either.

* Update build instructions with new method

It's just as simple to run conan and install the packages prior to the
cmake configure stage so lets now do that.

Move to using cmake presets to simply things and increase minimum
required version accordingly.

* Update circleCI config

Get the latest version of conan from pip, call conan externally prior
to cmake and use the autogenerated preset to configure and build.

* Update CI after changes to conan in version 2

Creating or detecting a new profile has changed and it is no longer
possible to edit settings from the command line. Luckily for us, it
appears that the setting we wanted to change to is now the default.

https://github.com/conan-io/conan/issues/12833

* Update CI config and README for older versions of cmake

Although presets arrived with 3.19, conan v2 uses functionality that was
release with 3.23. The current ubuntu LTS, which is used on circleCI,
has 3.22 so we can't use the presets created by conan in that config.

The presets do make config and build simpler, but we must assume that not
all people are on a new enough version of cmake so the README needs to
have instructions for those users.

* Move the conan step into cmake to simply configuration

New cmake file wraps up conan functionality and is called from the top
level cmake file. The setting of the toolchain file is done in the top
level file incase we decide to not do the conan step from within cmake.
If so, we can just delete the new file and the call to it.

The README and CI have also been updated to reflect these changes.

* More descriptive comment about the ignored file

Co-authored-by: ferdymercury <ferdymercury@users.noreply.github.com>

* Lower the necessary version of cmake

As we aren't yet using presets, there is no need to force the minimum version to be so high.

Co-authored-by: ferdymercury <ferdymercury@users.noreply.github.com>

* Update hash to latest qt-color-widgets commit

* Force non-interactive mode in the CI container

Ubuntu 22.04 appears to be interactive by default for apt processes so
we need to make it non-interactive so the CI builds can complete.

https://manpages.ubuntu.com/manpages/jammy/man1/needrestart.1.html

* Play around with option to get CI working

Apparently it's bad to `sudo export` and calls to `sudo` sanitise the
environment anyway. Use export as the standard user and then use the
`-E` option to preserve the environment variables. Combine the 3 install
commands into 1.

* Update spdlog package to 1.12.0

Remove specific dependency on fmt as that comes with spdlog and
having it separate was causing version conflicts.

* Add explicit reference to fmt back in

The fmt package is used independently to spdlog so we should explicitly
include it.

---------

Co-authored-by: ferdymercury <ferdymercury@users.noreply.github.com>