Put the unused parameters back in but mark as maybe_unused
It looks as if they were used in the past so might be used again so
for the moment, keep them in but use the maybe_unused attribute to
remove compilation warnings.

The logging function don't know how to format the Continuation struct
so leave the updated call commented for the moment.