Specify the build type in the conan command (#21)
* Specify the build type in the conan command

Make sure the build type for the conan installs is the same as that of
the top level, i.e. whatever the user sets when configuring this
project.

Start working on #20

* Forgot to set build type in CI script

* Build the CI pipeline in both Debug and Release builds

We can reuse config and just replace the build type. This may need a few
tweaks to get the syntax correct.

* Add missing colon to circleci config file

* Set a default build type in the CI config

* Ignore any directory that starts with 'build'

For testing purposes I have 2 build directories, build-debug and
build-release, I want git to ignore both.